Fix a bug whereby hardware-scaled PAL games could crash (GCW Zero)
If the video mode is set to 256x224 during game initialisation and the
SNES game sets the 239 height bit, formerly the video mode was not
changed to follow. However, 15 extra lines would be copied to the
screen, corrupting nearby memory.

A function called sal_VideoSetPAL was added to change the height during
gameplay.