Fix an oversight regarding PAL in commit aa75de9030e177620364300709bdf3ecc0f22dc5
As reported by hi-ban, the first frame following a reset or the loading
of a saved state modifies Memory.FillRAM outside of Run(int), in a way
that was not properly accounted for after sal_VideoEnterGame.

This fixes hardware-scaled 256x239 images crashing the emulator after
resets that cause the video mode to start out as 256x224.

Now the LastPAL variable is updated after the video mode is first
requested after exiting the menu.